Refactor BERT model implementation and remove redundant files
